* Code clean up * CvUnderTitle moved for separate component
